Move tiered bench tools into o11ytsdb bench
