+++ packages/o11ytsdb/bench/analyze-cpuprofile.mjs
