+++ packages/o11ytsdb/bench/analyze-heapprofile.mjs
