+++ packages/o11ytsdb/bench/chunk-size-sweep.ts
@@ -2,9 +2,9 @@ import { readFileSync } from "node:fs";
 import path from "node:path";
 import { performance } from "node:perf_hooks";
 
-import { ScanEngine } from "../packages/o11ytsdb/src/query.ts";
-import { RowGroupStore } from "../packages/o11ytsdb/src/row-group-store.ts";
-import { initWasmCodecs } from "../packages/o11ytsdb/src/wasm-codecs.ts";
+import { ScanEngine } from "../src/query.ts";
+import { RowGroupStore } from "../src/row-group-store.ts";
+import { initWasmCodecs } from "../src/wasm-codecs.ts";
 
 const NUM_SERIES = 32;
 const POINTS_PER_SERIES = 262_144;
