+++ packages/o11ytsdb/package.json
@@ -50,6 +50,10 @@
     "bench:ingest": "npm run bench -- ingest",
     "bench:query": "npm run bench -- query",
     "bench:rowgroup-layout": "npm run build && npm run bench:build && node --expose-gc bench/dist/rowgroup-layout.bench.js",
+    "bench:chunk-size-sweep": "npm run build && npm run bench:build && node bench/dist/chunk-size-sweep.js",
+    "bench:rowgroup-memory-audit": "npm run build && npm run bench:build && node bench/dist/rowgroup-memory-audit.js",
+    "bench:rowgroup-profile": "npm run build && npm run bench:build && node --expose-gc bench/dist/rowgroup-profile.js",
+    "bench:tiered-rowgroup": "npm run build && npm run bench:build && node bench/dist/tiered-rowgroup-compare.js",
     "bench:compare": "npm run bench -- --compare",
     "smoke": "npm run build && node bench/smoke.mjs",
     "typecheck": "tsc -p tsconfig.json --noEmit"
